refactor the game controls into its own file for readability
