+++ app/alien_invasion.py
@@ -1,9 +1,10 @@
 import pygame
 from pygame.sprite import Group
 
+import app.game_functions as gf
+import app.game_controls as gc
 from app.settings import Settings
 from app.ship import Ship
-import app.game_functions as gf
 from app.game_stats import GameStats
 from app.button import Button
 from app.scoreboard import Scoreboard
@@ -31,7 +32,7 @@ def run_game():
 
     # Start the main loop for the game
     while True:
-        gf.check_events(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets)
+        gc.check_events(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets)
 
         if stats.game_active:
             ship.update()
