+++ app/game_controls.py
@@ -1,0 +1,43 @@
+import sys
+
+import pygame
+
+import app.game_functions as gf
+
+
+# Event handling
+def check_events(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets):
+    """Respond to keypresses and mouse events."""
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            sys.exit()
+        elif event.type == pygame.MOUSEBUTTONDOWN:
+            mouse_x, mouse_y = pygame.mouse.get_pos()
+            if not stats.game_active:
+                gf.check_play_button(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets, mouse_x, mouse_y)
+        # Ship movement
+        elif event.type == pygame.KEYDOWN:
+            check_keydown_events(event, ai_settings, stats, screen, ship, bullets)
+        elif event.type == pygame.KEYUP:
+            check_keyup_events(event, ship)
+
+
+def check_keydown_events(event, ai_settings, stats, screen, ship, bullets):
+    """Respond to keypresses."""
+    if event.key == pygame.K_RIGHT:
+        ship.moving_right = True
+    elif event.key == pygame.K_LEFT:
+        ship.moving_left = True
+    elif event.key == pygame.K_SPACE:
+        gf.fire_bullet(ai_settings, bullets, screen, ship)
+    elif event.key == pygame.K_p:
+        stats.game_active = True
+    elif event.key == pygame.K_q:
+        sys.exit()
+
+
+def check_keyup_events(event, ship):
+    if event.key == pygame.K_RIGHT:
+        ship.moving_right = False
+    elif event.key == pygame.K_LEFT:
+        ship.moving_left = False
