+++ app/game_functions.py
@@ -29,42 +29,42 @@ def check_play_button(ai_settings, screen, stats, sb, play_button, ship, aliens,
         pygame.mouse.set_visible(True)
 
 
-# Event handling
-def check_events(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets):
-    """Respond to keypresses and mouse events."""
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT:
-            sys.exit()
-        elif event.type == pygame.MOUSEBUTTONDOWN:
-            mouse_x, mouse_y = pygame.mouse.get_pos()
-            if not stats.game_active:
-                check_play_button(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets, mouse_x, mouse_y)
-        # Ship movement
-        elif event.type == pygame.KEYDOWN:
-            check_keydown_events(event, ai_settings, stats, screen, ship, bullets)
-        elif event.type == pygame.KEYUP:
-            check_keyup_events(event, ship)
-
-
-def check_keydown_events(event, ai_settings, stats, screen, ship, bullets):
-    """Respond to keypresses."""
-    if event.key == pygame.K_RIGHT:
-        ship.moving_right = True
-    elif event.key == pygame.K_LEFT:
-        ship.moving_left = True
-    elif event.key == pygame.K_SPACE:
-        fire_bullet(ai_settings, bullets, screen, ship)
-    elif event.key == pygame.K_p:
-        stats.game_active = True
-    elif event.key == pygame.K_q:
-        sys.exit()
-
-
-def check_keyup_events(event, ship):
-    if event.key == pygame.K_RIGHT:
-        ship.moving_right = False
-    elif event.key == pygame.K_LEFT:
-        ship.moving_left = False
+# # Event handling
+# def check_events(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets):
+#     """Respond to keypresses and mouse events."""
+#     for event in pygame.event.get():
+#         if event.type == pygame.QUIT:
+#             sys.exit()
+#         elif event.type == pygame.MOUSEBUTTONDOWN:
+#             mouse_x, mouse_y = pygame.mouse.get_pos()
+#             if not stats.game_active:
+#                 check_play_button(ai_settings, screen, stats, sb, play_button, ship, aliens, bullets, mouse_x, mouse_y)
+#         # Ship movement
+#         elif event.type == pygame.KEYDOWN:
+#             check_keydown_events(event, ai_settings, stats, screen, ship, bullets)
+#         elif event.type == pygame.KEYUP:
+#             check_keyup_events(event, ship)
+#
+#
+# def check_keydown_events(event, ai_settings, stats, screen, ship, bullets):
+#     """Respond to keypresses."""
+#     if event.key == pygame.K_RIGHT:
+#         ship.moving_right = True
+#     elif event.key == pygame.K_LEFT:
+#         ship.moving_left = True
+#     elif event.key == pygame.K_SPACE:
+#         fire_bullet(ai_settings, bullets, screen, ship)
+#     elif event.key == pygame.K_p:
+#         stats.game_active = True
+#     elif event.key == pygame.K_q:
+#         sys.exit()
+#
+#
+# def check_keyup_events(event, ship):
+#     if event.key == pygame.K_RIGHT:
+#         ship.moving_right = False
+#     elif event.key == pygame.K_LEFT:
+#         ship.moving_left = False
 
 
 # Bullet handling
@@ -90,16 +90,23 @@ def check_bullet_alien_collisions(ai_settings, screen, stats, sb, ship, aliens, 
     collisions = pygame.sprite.groupcollide(bullets, aliens, True, True)
     if collisions:
         for aliens in collisions.values():
-            stats.score += ai_settings.alien_points * len(aliens)
-            sb.prep_score()
+            increase_score(ai_settings, aliens, sb, stats)
         check_high_score(stats, sb)
     if len(aliens) == 0:
-        bullets.empty()
-        ai_settings.increase_speed()
-        stats.level += 1
-        sb.prep_level()
+        start_new_level(ai_settings, aliens, bullets, sb, screen, ship, stats)
 
-        create_fleet(ai_settings, screen, ship, aliens)
+
+def increase_score(ai_settings, aliens, sb, stats):
+    stats.score += ai_settings.alien_points * len(aliens)
+    sb.prep_score()
+
+
+def start_new_level(ai_settings, aliens, bullets, sb, screen, ship, stats):
+    bullets.empty()
+    ai_settings.increase_speed()
+    stats.level += 1
+    sb.prep_level()
+    create_fleet(ai_settings, screen, ship, aliens)
 
 
 def check_high_score(stats, sb):
@@ -190,7 +197,6 @@ def update_aliens(ai_settings, stats, screen, sb, ship, aliens, bullets):
 
     if pygame.sprite.spritecollideany(ship, aliens):
         ship_hit(ai_settings, stats, screen, sb, ship, aliens, bullets)
-        print("Ship hit!!!")
 
 
 # Screen handling
